% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_feature_dataset.R
\name{build_feature_dataset}
\alias{build_feature_dataset}
\title{Build feature set for machine learning model}
\usage{
build_feature_dataset(data, mol, features, target, data_path)
}
\arguments{
\item{mol}{Molecular signatures from untreated sample}

\item{features}{drug_name, chemical_structure, molecular (top125), geneset (top125)}

\item{target}{'aoc' or bliss'}
}
\value{

}
\description{
Build feature set for machine learning model
}
