% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_optimal_drug_combination.R
\name{predict_optimal_drug_combination}
\alias{predict_optimal_drug_combination}
\title{Predict best drug combinations based on patient's molecular readouts}
\usage{
predict_optimal_drug_combination(mol_df)
}
\arguments{
\item{mol_df}{data.frame with molecualr readouts from patient}
}
\value{
best drug combinations based on prediction
}
\description{
Predict best drug combinations based on patient's molecular readouts
}
